KEP-2676: Support Istio ambient in Kubeflow
Add proposal for supporting Istio ambient in Kubeflow. This has been
discussed for more than a year in the Kubeflow community (see #2676).
In order for Kubeflow users to benefit from the reduced resource
footprint and simplified operations in ambient mode, this proposal
defines the changes needed to make Kubeflow ambient-compatible.

Ref #2676